+++ appender/src/main/java/com/vlkan/log4j2/redis/appender/RedisThrottler.java
@@ -3,6 +3,7 @@ package com.vlkan.log4j2.redis.appender;
 import com.vlkan.log4j2.redis.appender.guava.GuavaRateLimiter;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.core.LoggerContext;
+import org.apache.logging.log4j.status.StatusLogger;
 
 import javax.management.InstanceAlreadyExistsException;
 import javax.management.JMX;
@@ -21,19 +22,25 @@ import java.util.concurrent.atomic.AtomicReference;
 
 class RedisThrottler implements AutoCloseable {
 
+    private static final StatusLogger LOGGER = StatusLogger.getLogger();
+
     /**
      * Reference count of JMX beans.
      * <p>
-     * Certain applications (e.g., Spring Boot) known to reconfigure <code>LoggerContext</code> multiple times. This
-     * triggers multiple interleaved start-stop calls causing <code>RedisThrottler</code> to unregister an in-use
-     * JMX bean. <code>jmxBeanReferenceCountByName</code> keeps the reference counts to created JMX beans and
-     * unregisters them at close if there are no more references.
+     * Certain applications (e.g., Spring Boot) known to reconfigure
+     * <code>LoggerContext</code> multiple times. This triggers multiple
+     * interleaved start-stop calls causing <code>RedisThrottler</code> to
+     * unregister an in-use JMX bean. This map keeps the reference counts to
+     * created JMX beans and unregisters them at close if there are no more
+     * references.
      */
-    private static final Map<ObjectName, Integer> jmxBeanReferenceCountByName = new HashMap<>();
+    private static final Map<ObjectName, Integer> JMX_BEAN_REFERENCE_COUNT_BY_NAME = new HashMap<>();
 
     private final RedisThrottlerConfig config;
 
     private final RedisAppender appender;
+
+    private final String logPrefix;
 
     private final boolean ignoreExceptions;
 
@@ -47,28 +54,27 @@ class RedisThrottler implements AutoCloseable {
 
     private final GuavaRateLimiter byteRateLimiter;
 
-    private final DebugLogger logger;
-
     private final ObjectName jmxBeanName;
+
+    private volatile boolean started = false;
 
     private volatile RedisThrottlerJmxBean jmxBean = null;
 
-    private final AtomicReference<Throwable> lastThrown = new AtomicReference<>(null);
+    private final AtomicReference<Throwable> lastThrownRef = new AtomicReference<>(null);
 
     RedisThrottler(
             RedisThrottlerConfig config,
             RedisAppender appender,
-            boolean ignoreExceptions,
-            boolean debugEnabled) {
+            boolean ignoreExceptions) {
         this.config = config;
         this.appender = appender;
+        this.logPrefix = String.format("[RedisThrottler{%s}]", appender.getName());
         this.ignoreExceptions = ignoreExceptions;
         this.buffer = new ArrayBlockingQueue<>(config.getBufferSize());
         this.batch = new byte[config.getBatchSize()][];
-        this.flushTrigger = createFlushTrigger();
+        this.flushTrigger = createFlushTrigger(appender.getName());
         this.eventRateLimiter = config.getMaxEventCountPerSecond() > 0 ? GuavaRateLimiter.create(config.getMaxEventCountPerSecond()) : null;
         this.byteRateLimiter = config.getMaxByteCountPerSecond() > 0 ? GuavaRateLimiter.create(config.getMaxByteCountPerSecond()) : null;
-        this.logger = new DebugLogger(RedisThrottler.class, debugEnabled);
         this.jmxBeanName = createJmxBeanName();
     }
 
@@ -92,89 +98,111 @@ class RedisThrottler implements AutoCloseable {
         }
     }
 
-    private Thread createFlushTrigger() {
+    private Thread createFlushTrigger(String appenderName) {
         Thread thread = new Thread(this::flushContinuously);
+        thread.setName(appenderName + " Throttler");
         thread.setDaemon(true);
         return thread;
     }
 
     private void flushContinuously() {
-        logger.debug("started");
-        do {
-            logger.debug("flushing");
+
+        // Determine the wait period.
+        long waitPeriodNanos = Math.multiplyExact(1_000_000L, config.getFlushPeriodMillis());
+        if (LOGGER.isInfoEnabled()) {
+            String waitPeriod = String.format("%.3fs", waitPeriodNanos * 1e-9);
+            LOGGER.info("{} started (waitPeriod={})", logPrefix, waitPeriod);
+        }
+
+        // Flush continuously.
+        while (started) {
+            LOGGER.debug("{} flushing", logPrefix);
             try {
-                flush();
+                flush(waitPeriodNanos);
             } catch (InterruptedException ignored) {
-                logger.debug("interrupted");
+                LOGGER.debug("{} interrupted", logPrefix);
                 Thread.currentThread().interrupt();
-                break;
             }
-        } while (true);
+        }
+
+        // Flush one last time for any left overs in the buffer.
+        LOGGER.debug("{} flushing one last time", logPrefix);
+        try {
+            flush(waitPeriodNanos);
+        } catch (InterruptedException ignored) {
+            LOGGER.debug("{} interrupted", logPrefix);
+            Thread.currentThread().interrupt();
+        }
+
     }
 
-    private void flush() throws InterruptedException {
+    private void flush(long waitPeriodNanos) throws InterruptedException {
 
         int batchIndex = 0;
         byte[] event;
 
         // Flush in batches.
-        logger.debug("polling");
-        long waitPeriodMillis = config.getFlushPeriodMillis();
-        while (waitPeriodMillis > 0) {
-            long pollTimeMillis = System.currentTimeMillis();
-            event = buffer.poll(waitPeriodMillis, TimeUnit.MILLISECONDS);
+        LOGGER.debug("{} polling", logPrefix);
+        while (waitPeriodNanos > 0) {
+            long pollTimeNanos = System.nanoTime();
+            event = buffer.poll(waitPeriodNanos, TimeUnit.NANOSECONDS);
             if (event == null) {
                 break;
             }
-            if (logger.isEnabled()) {
-                logger.debug("polled: %s", new String(event));
+            if (LOGGER.isTraceEnabled()) {
+                LOGGER.trace("{} polled: {}", logPrefix, new String(event));
             }
             batch[batchIndex++] = event;
             if (batchIndex == batch.length) {
-                safeConsumeEvents(batch);
+                push(batch);
                 batchIndex = 0;
             }
-            long pollPeriodMillis = System.currentTimeMillis() - pollTimeMillis;
-            waitPeriodMillis -= pollPeriodMillis;
+            long pollPeriodNanos = System.nanoTime() - pollTimeNanos;
+            waitPeriodNanos -= pollPeriodNanos;
         }
 
         // Flush the last remaining.
         if (batchIndex > 0) {
-            logger.debug("pushing remaining %d events", batchIndex);
+            LOGGER.debug("{} pushing remaining {} events", logPrefix, batchIndex);
             byte[][] subBatch = Arrays.copyOfRange(batch, 0, batchIndex);
-            safeConsumeEvents(subBatch);
+            push(subBatch);
         }
 
     }
 
-    private void safeConsumeEvents(final byte[]... events) {
+    private void push(final byte[][] events) {
         int eventCount = events.length;
         try {
-            logger.debug("pushing %d events", eventCount);
+            LOGGER.debug("{} pushing {} events", logPrefix, eventCount);
             appender.consumeThrottledEvents(events);
             jmxBean.incrementRedisPushSuccessCount(eventCount);
-        } catch (Throwable thrown) {
-            if (logger.isEnabled()) {
-                logger.debug("push failure: %s", thrown.getMessage());
+        } catch (Exception thrown) {
+            if (LOGGER.isWarnEnabled()) {
+                LOGGER.warn("{} push failure", logPrefix, thrown);
                 thrown.printStackTrace();
             }
-            lastThrown.set(thrown);
+            lastThrownRef.set(thrown);
             jmxBean.incrementRedisPushFailureCount(eventCount);
         }
     }
 
-    public RedisThrottlerJmxBean getJmxBean() {
+    RedisThrottlerJmxBean getJmxBean() {
         return jmxBean;
     }
 
-    public void push(byte[] event) {
+    synchronized void push(byte[] event) {
+
+        if (!started) {
+            LOGGER.debug("{} not started yet, ignoring the push request", logPrefix);
+            return;
+        }
 
         jmxBean.incrementTotalEventCount(1);
 
-        Throwable thrown = lastThrown.getAndSet(null);
-        if (thrown != null) {
+        Throwable lastThrown = lastThrownRef.getAndSet(null);
+        if (lastThrown != null) {
             jmxBean.incrementIgnoredEventCount(1);
-            tryThrow(thrown);
+            tryThrow("failed pushing due to an earlier throttler failure", lastThrown);
             return;
         }
 
@@ -197,31 +225,37 @@ class RedisThrottler implements AutoCloseable {
 
     }
 
-    private void tryThrow(Throwable error) {
-        logger.debug(error.getMessage(), error);
+    @SuppressWarnings("SameParameterValue")
+    private void tryThrow(String message, Throwable error) {
+        if (LOGGER.isErrorEnabled()) {
+            LOGGER.error(logPrefix + " " + message, error);
+        }
         if (!ignoreExceptions)
             throw new RuntimeException(error);
     }
 
     private void tryThrow(String error) {
-        logger.debug(error);
+        LOGGER.error(logPrefix + " " + error);
         if (!ignoreExceptions)
             throw new RuntimeException(error);
     }
 
-    public synchronized void start() {
-        logger.debug("starting");
-        jmxBean = registerOrGetJmxBean();
-        flushTrigger.start();
+    synchronized void start() {
+        if (!started) {
+            LOGGER.info("{} starting", logPrefix);
+            started = true;
+            jmxBean = registerOrGetJmxBean();
+            flushTrigger.start();
+        }
     }
 
     private RedisThrottlerJmxBean registerOrGetJmxBean() {
         MBeanServer mbs = ManagementFactory.getPlatformMBeanServer();
         try {
-            synchronized (jmxBeanReferenceCountByName) {
+            synchronized (JMX_BEAN_REFERENCE_COUNT_BY_NAME) {
 
                 // Get the reference count for the JMX bean.
-                Integer jmxBeanReferenceCount = jmxBeanReferenceCountByName.get(jmxBeanName);
+                Integer jmxBeanReferenceCount = JMX_BEAN_REFERENCE_COUNT_BY_NAME.get(jmxBeanName);
                 if (jmxBeanReferenceCount == null) {
                     jmxBeanReferenceCount = 0;
                 }
@@ -237,11 +271,11 @@ class RedisThrottler implements AutoCloseable {
                 }
 
                 // Increment the reference count and return the JMX bean.
-                jmxBeanReferenceCountByName.put(jmxBeanName, jmxBeanReferenceCount + 1);
+                JMX_BEAN_REFERENCE_COUNT_BY_NAME.put(jmxBeanName, jmxBeanReferenceCount + 1);
                 return jmxBean;
 
             }
-        } catch (Throwable error) {
+        } catch (Exception error) {
             String message = String.format("failed accessing the JMX bean (jmxBeanName=%s)", jmxBeanName);
             throw new RuntimeException(message, error);
         }
@@ -249,38 +283,44 @@ class RedisThrottler implements AutoCloseable {
 
     @Override
     public synchronized void close() {
-        logger.debug("closing");
-        flushTrigger.interrupt();
-        unregisterJmxBean();
+        if (started) {
+            LOGGER.info("{} closing", logPrefix);
+            started = false;
+            flushTrigger.interrupt();
+            unregisterJmxBean();
+        }
     }
 
     private void unregisterJmxBean() {
         MBeanServer mbs = ManagementFactory.getPlatformMBeanServer();
-        synchronized (jmxBeanReferenceCountByName) {
+        synchronized (JMX_BEAN_REFERENCE_COUNT_BY_NAME) {
 
             // Get the reference count for the JMX bean.
-            Integer jmxBeanReferenceCount = jmxBeanReferenceCountByName.get(jmxBeanName);
+            Integer jmxBeanReferenceCount = JMX_BEAN_REFERENCE_COUNT_BY_NAME.get(jmxBeanName);
 
             // Check if we have a valid state, that is, jmxBeanReferenceCount > 0.
             if (jmxBeanReferenceCount == null || jmxBeanReferenceCount == 0) {
-                logger.debug(
-                        "failed unregistering the JMX bean (jmxBeanName=%s, jmxBeanReferenceCount=%s)",
-                        jmxBeanName, jmxBeanReferenceCount);
+                LOGGER.warn(
+                        "{} failed unregistering the JMX bean (jmxBeanName={}, jmxBeanReferenceCount={})",
+                        logPrefix, jmxBeanName, jmxBeanReferenceCount);
             }
 
             // If there is just a single reference so far, it is safe to unregister the bean.
             else if (jmxBeanReferenceCount == 1) {
                 try {
                     mbs.unregisterMBean(jmxBeanName);
-                    jmxBeanReferenceCountByName.remove(jmxBeanName);
-                } catch (Throwable error) {
-                    logger.debug("failed unregistering the JMX bean (jmxBeanName=%s)", jmxBeanName);
+                    JMX_BEAN_REFERENCE_COUNT_BY_NAME.remove(jmxBeanName);
+                } catch (Exception error) {
+                    String message = String.format(
+                            "%s failed unregistering the JMX bean (jmxBeanName=%s)",
+                            logPrefix, jmxBeanName);
+                    LOGGER.error(message, error);
                 }
             }
 
             // Apparently there are more consumers of the bean. Just decrement the reference count.
             else {
-                jmxBeanReferenceCountByName.put(jmxBeanName, jmxBeanReferenceCount - 1);
+                JMX_BEAN_REFERENCE_COUNT_BY_NAME.put(jmxBeanName, jmxBeanReferenceCount - 1);
             }
 
         }