+++ appender/src/test/java/com/vlkan/log4j2/redis/appender/RedisAppenderTestConfig.java
@@ -11,11 +11,11 @@ enum RedisAppenderTestConfig {;
 
     static final String REDIS_HOST = "localhost";
 
-    static final String REDIS_PASSWORD = "toosecret";
+    static final String REDIS_PASSWORD = "milonga";
 
     static final int REDIS_PORT = 63790;
 
-    static final URI LOG4J2_CONFIG_FILE_URI = TestHelpers.uri("classpath:log4j2.RedisAppenderTest.xml");
+    static final URI LOG4J2_CONFIG_FILE_URI = URI.create("classpath:log4j2.RedisAppenderTest.xml");
 
     static final String LOG4J2_APPENDER_NAME = "REDIS";