+++ appender/src/test/java/com/vlkan/log4j2/redis/appender/RedisTestConstants.java
@@ -1,0 +1,9 @@
+package com.vlkan.log4j2.redis.appender;
+
+import java.util.Random;
+
+enum RedisTestConstants {;
+
+    static final Random RANDOM = new Random(0);
+
+}